+++ src/logging.rs
@@ -2,6 +2,7 @@ use std::env;
 
 extern crate pretty_env_logger;
 
+/// Initialize the logger
 pub fn init() {
     // Set default log level if none is set
     if env::var("RUST_LOG").is_err() {